Fix: Ensure all or none Schema::create methods execute in single migration file with multiple calls
This resolves an issue where migrations containing multiple Schema::create
method calls would only partially execute. The issue was caused by
incomplete handling of Schema operations in a single migration file. Initial schemas defined inside
up method of migration file would be created in database without logging that migration file in
migrations repository if there were some errors in subsequent Schema::create calls.
The real issue comes when running php artisan migrate command second time. It will throw table already
exists exception preventing the migration file from creating other schemas specified after the schema with
error.

This fix ensures that, if Schema::create method inside up method of migration file is partially run, it runs
corresponding down method to drop those schemas if any error occours in subsequent create methods of Schema
class in up method of the same migration file.

+++ src/Illuminate/Database/Migrations/DatabaseMigrationRepository.php
@@ -130,7 +130,8 @@ class DatabaseMigrationRepository implements MigrationRepositoryInterface
      */
     public function delete($migration)
     {
-        $this->table()->where('migration', $migration->migration)->delete();
+        if(property_exists($migration, 'migration'))
+            $this->table()->where('migration', $migration->migration)->delete();
     }
 
     /**
